backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic>=2.8.0
python-dotenv==1.0.0
httpx>=0.27
bcrypt==4.1.2
PyJWT==2.8.0
